+++ src/main/java/konkuk/thip/user/application/service/following/UserFollowService.java
@@ -1,4 +1,4 @@
-package konkuk.thip.user.application.service.follow;
+package konkuk.thip.user.application.service.following;
 
 import konkuk.thip.common.exception.InvalidStateException;
 import konkuk.thip.user.application.port.in.UserFollowUsecase;
